refactor(db): move event query store ownership
