+++ crates/buzz-db/src/lib.rs
@@ -1181,23 +1181,6 @@ impl Db {
             .map_err(Into::into)
     }
 
-    /// Inserts an event. Returns `(StoredEvent, was_inserted)` — `false` on duplicate.
-    #[datastore_span(name = "insert_event", system = "postgresql")]
-    pub async fn insert_event(
-        &self,
-        community_id: CommunityId,
-        event: &nostr::Event,
-        channel_id: Option<Uuid>,
-    ) -> Result<(StoredEvent, bool)> {
-        let result = event::insert_event(&self.pool, community_id, event, channel_id).await?;
-        if result.1 {
-            if let Err(e) = insert_mentions(&self.pool, community_id, event, channel_id).await {
-                tracing::warn!(event_id = %event.id, "Failed to insert mentions: {e}");
-            }
-        }
-        Ok(result)
-    }
-
     /// Insert an event while holding and validating an admitted serving-write
     /// lease under the community ordering lock through commit.
     ///
@@ -1239,304 +1222,6 @@ impl Db {
             }
         }
         Ok(result)
-    }
-
-    /// Queries events matching the given filter parameters.
-    ///
-    /// Always reads from the WRITER pool. If the result influences a write
-    /// or a permission decision, this is the method to call. Display-path
-    /// callers that tolerate bounded staleness should use
-    /// [`Db::query_events_routed`] instead — converting a caller is an
-    /// explicit, per-callsite decision, never a change to this method.
-    #[datastore_span(name = "query_events", system = "postgresql")]
-    pub async fn query_events(&self, q: &EventQuery) -> Result<Vec<StoredEvent>> {
-        event::query_events(&self.pool, q).await
-    }
-
-    /// [`Db::query_events`] with replica routing — the opt-in fast path for
-    /// display reads.
-    ///
-    /// Rule of thumb: **if the result influences a write or a permission,
-    /// it reads from the writer** — do not convert such a caller to this
-    /// method. Every new caller must be added to the caller-classification
-    /// table in `PLANS/REPLICA_FULL_READ_ROUTING_DESIGN.md`.
-    ///
-    /// Routing derives the strongest sound predicate from the query shape
-    /// ([`RoutePredicate::for_query`]): a channel-pinned query with an
-    /// `until` upper bound may be served covered (provably complete below
-    /// the fence wall); anything else is bounded-staleness only. The whole
-    /// seam is gated on `BUZZ_REPLICA_READ_MAX_AGE_MS` (default off): when
-    /// unset, even covered-eligible queries stay on the writer, so merging
-    /// this seam is a true no-op until the budget is configured. Every
-    /// failure fails closed to the writer.
-    #[datastore_span(name = "query_events_routed", system = "postgresql")]
-    pub async fn query_events_routed(
-        &self,
-        path: &'static str,
-        q: &EventQuery,
-    ) -> Result<Vec<StoredEvent>> {
-        let predicate = RoutePredicate::for_query(q, self.replica_read_max_age.is_some());
-        match self.route_read(path, predicate).await {
-            RouteDecision::Replica(mut tx, _entry, reason) => {
-                match event::query_events_on(&mut tx, q).await {
-                    Ok(events) => {
-                        Self::record_route(path, "replica", reason);
-                        Ok(events)
-                    }
-                    Err(e) => {
-                        // Mid-query replica failure: fail closed to the
-                        // writer rather than surfacing a routed error.
-                        tracing::warn!(path, "replica read failed; re-running on writer: {e}");
-                        Self::record_route(path, "writer", "replica_error");
-                        event::query_events(&self.pool, q).await
-                    }
-                }
-            }
-            RouteDecision::Writer => event::query_events(&self.pool, q).await,
-        }
-    }
-
-    /// [`Db::query_events_routed`] restricted to the BOUNDED arm — for
-    /// reads whose result feeds a COUNT rather than a displayed page.
-    ///
-    /// The covered arm bounds insert-completeness only; stale deletions can
-    /// briefly inflate the result set (see [`RoutePredicate::Covered`]). A
-    /// display page absorbs that per-row; a number derived from the rows
-    /// does not. Same classification-table requirement as
-    /// [`Db::query_events_routed`].
-    #[datastore_span(name = "query_events_routed_bounded", system = "postgresql")]
-    pub async fn query_events_routed_bounded(
-        &self,
-        path: &'static str,
-        q: &EventQuery,
-    ) -> Result<Vec<StoredEvent>> {
-        match self.route_read(path, RoutePredicate::Bounded).await {
-            RouteDecision::Replica(mut tx, _entry, reason) => {
-                match event::query_events_on(&mut tx, q).await {
-                    Ok(events) => {
-                        Self::record_route(path, "replica", reason);
-                        Ok(events)
-                    }
-                    Err(e) => {
-                        tracing::warn!(path, "replica read failed; re-running on writer: {e}");
-                        Self::record_route(path, "writer", "replica_error");
-                        event::query_events(&self.pool, q).await
-                    }
-                }
-            }
-            RouteDecision::Writer => event::query_events(&self.pool, q).await,
-        }
-    }
-
-    /// Count events matching the given query (NIP-45 COUNT support).
-    ///
-    /// Always reads from the WRITER pool — see [`Db::query_events`] for the
-    /// writer-vs-routed rule.
-    #[datastore_span(name = "count_events", system = "postgresql")]
-    pub async fn count_events(&self, q: &EventQuery) -> Result<i64> {
-        event::count_events(&self.pool, q).await
-    }
-
-    /// [`Db::count_events`] with replica routing — same contract, rules,
-    /// and classification-table requirement as [`Db::query_events_routed`].
-    ///
-    /// Counts route on the BOUNDED arm only, never covered: the covered
-    /// arm bounds insert-completeness but not deletion visibility (soft
-    /// deletes are UPDATEs outside the floor guard), and a count has no
-    /// downstream per-row re-filter to absorb extra rows — a silently
-    /// inflated number for up to `FENCE_STALENESS` is a different product
-    /// statement than a page briefly showing a deleted row. `Bounded` ties
-    /// the error to the accepted budget `B`.
-    #[datastore_span(name = "count_events_routed", system = "postgresql")]
-    pub async fn count_events_routed(&self, path: &'static str, q: &EventQuery) -> Result<i64> {
-        match self.route_read(path, RoutePredicate::Bounded).await {
-            RouteDecision::Replica(mut tx, _entry, reason) => {
-                match event::count_events_on(&mut tx, q).await {
-                    Ok(count) => {
-                        Self::record_route(path, "replica", reason);
-                        Ok(count)
-                    }
-                    Err(e) => {
-                        tracing::warn!(path, "replica count failed; re-running on writer: {e}");
-                        Self::record_route(path, "writer", "replica_error");
-                        event::count_events(&self.pool, q).await
-                    }
-                }
-            }
-            RouteDecision::Writer => event::count_events(&self.pool, q).await,
-        }
-    }
-
-    /// Return whether a creator-signed huddle-start event links a parent
-    /// channel to an ephemeral huddle channel.
-    #[datastore_span(name = "huddle_started_link_exists", system = "postgresql")]
-    pub async fn huddle_started_link_exists(
-        &self,
-        community_id: CommunityId,
-        parent_channel_id: Uuid,
-        ephemeral_channel_id: Uuid,
-        creator_pubkey: &[u8],
-    ) -> Result<bool> {
-        event::huddle_started_link_exists(
-            &self.pool,
-            community_id,
-            parent_channel_id,
-            ephemeral_channel_id,
-            creator_pubkey,
-        )
-        .await
-    }
-
-    /// Fetch the latest replaceable event for a (kind, pubkey) pair.
-    ///
-    /// Uses canonical NIP-16 ordering: `created_at DESC, id ASC`.
-    /// This matches the write path in [`replace_addressable_event`] and handles
-    /// historical duplicate survivors correctly.
-    #[datastore_span(name = "get_latest_global_replaceable", system = "postgresql")]
-    pub async fn get_latest_global_replaceable(
-        &self,
-        community_id: CommunityId,
-        kind: i32,
-        pubkey_bytes: &[u8],
-    ) -> Result<Option<StoredEvent>> {
-        event::get_latest_global_replaceable(&self.pool, community_id, kind, pubkey_bytes).await
-    }
-
-    /// Fetches a single non-deleted event by its raw ID bytes.
-    ///
-    /// Returns `None` if the event does not exist or has been soft-deleted.
-    #[datastore_span(name = "get_event_by_id", system = "postgresql")]
-    pub async fn get_event_by_id(
-        &self,
-        community_id: CommunityId,
-        id_bytes: &[u8],
-    ) -> Result<Option<StoredEvent>> {
-        event::get_event_by_id(&self.pool, community_id, id_bytes).await
-    }
-
-    /// Fetches a single event by its raw ID bytes, **including soft-deleted rows**.
-    #[datastore_span(name = "get_event_by_id_including_deleted", system = "postgresql")]
-    pub async fn get_event_by_id_including_deleted(
-        &self,
-        community_id: CommunityId,
-        id_bytes: &[u8],
-    ) -> Result<Option<StoredEvent>> {
-        event::get_event_by_id_including_deleted(&self.pool, community_id, id_bytes).await
-    }
-
-    /// Soft-deletes an event. Returns `Ok(true)` if deleted, `Ok(false)` if already deleted.
-    #[datastore_span(name = "soft_delete_event", system = "postgresql")]
-    pub async fn soft_delete_event(
-        &self,
-        community_id: CommunityId,
-        event_id: &[u8],
-    ) -> Result<bool> {
-        event::soft_delete_event(&self.pool, community_id, event_id).await
-    }
-
-    /// Soft-delete the live row for an addressable coordinate `(kind, pubkey, d_tag)`
-    /// when it is not newer than the deletion request.
-    /// Used by NIP-09 a-tag deletion for parameterized-replaceable kinds;
-    /// `deletion_created_at_secs` is the deletion event's `created_at`.
-    #[datastore_span(name = "soft_delete_by_coordinate", system = "postgresql")]
-    pub async fn soft_delete_by_coordinate(
-        &self,
-        community_id: CommunityId,
-        kind: i32,
-        pubkey: &[u8],
-        d_tag: &str,
-        deletion_created_at_secs: i64,
-    ) -> Result<bool> {
-        event::soft_delete_by_coordinate(
-            &self.pool,
-            community_id,
-            kind,
-            pubkey,
-            d_tag,
-            deletion_created_at_secs,
-        )
-        .await
-    }
-
-    /// Atomically soft-delete an event and decrement thread reply counters.
-    #[datastore_span(name = "soft_delete_event_and_update_thread", system = "postgresql")]
-    pub async fn soft_delete_event_and_update_thread(
-        &self,
-        community_id: CommunityId,
-        event_id: &[u8],
-        parent_event_id: Option<&[u8]>,
-        root_event_id: Option<&[u8]>,
-    ) -> Result<bool> {
-        event::soft_delete_event_and_update_thread(
-            &self.pool,
-            community_id,
-            event_id,
-            parent_event_id,
-            root_event_id,
-        )
-        .await
-    }
-
-    /// Returns the most recent `created_at` for a channel.
-    #[datastore_span(name = "get_last_message_at", system = "postgresql")]
-    pub async fn get_last_message_at(
-        &self,
-        community_id: CommunityId,
-        channel_id: Uuid,
-    ) -> Result<Option<DateTime<Utc>>> {
-        event::get_last_message_at(&self.pool, community_id, channel_id).await
-    }
-
-    /// Bulk-fetch the most recent `created_at` for a set of channel IDs.
-    #[datastore_span(name = "get_last_message_at_bulk", system = "postgresql")]
-    pub async fn get_last_message_at_bulk(
-        &self,
-        community_id: CommunityId,
-        channel_ids: &[Uuid],
-    ) -> Result<std::collections::HashMap<Uuid, DateTime<Utc>>> {
-        event::get_last_message_at_bulk(&self.pool, community_id, channel_ids).await
-    }
-
-    /// Batch-fetch non-deleted events by their raw IDs.
-    #[datastore_span(name = "get_events_by_ids", system = "postgresql")]
-    pub async fn get_events_by_ids(
-        &self,
-        community_id: CommunityId,
-        ids: &[&[u8]],
-    ) -> Result<Vec<StoredEvent>> {
-        event::get_events_by_ids(&self.pool, community_id, ids).await
-    }
-
-    /// [`Db::get_events_by_ids`] with replica routing — same contract and
-    /// classification-table requirement as [`Db::query_events_routed`].
-    ///
-    /// By-id fetches route on the BOUNDED arm only: an id list carries no
-    /// channel pin, so no fence floor can prove insert-completeness — the
-    /// covered arm is structurally unavailable. Used for FTS hit hydration,
-    /// where a missing row degrades to a skipped search hit downstream.
-    #[datastore_span(name = "get_events_by_ids_routed", system = "postgresql")]
-    pub async fn get_events_by_ids_routed(
-        &self,
-        path: &'static str,
-        community_id: CommunityId,
-        ids: &[&[u8]],
-    ) -> Result<Vec<StoredEvent>> {
-        match self.route_read(path, RoutePredicate::Bounded).await {
-            RouteDecision::Replica(mut tx, _entry, reason) => {
-                match event::get_events_by_ids_on(&mut tx, community_id, ids).await {
-                    Ok(events) => {
-                        Self::record_route(path, "replica", reason);
-                        Ok(events)
-                    }
-                    Err(e) => {
-                        tracing::warn!(path, "replica read failed; re-running on writer: {e}");
-                        Self::record_route(path, "writer", "replica_error");
-                        event::get_events_by_ids(&self.pool, community_id, ids).await
-                    }
-                }
-            }
-            RouteDecision::Writer => event::get_events_by_ids(&self.pool, community_id, ids).await,
-        }
     }
 
     /// Exclusively claim a batch of due matcher jobs from one community.
@@ -1706,31 +1391,6 @@ impl Db {
             max_active_leases,
         )
         .await
-    }
-
-    /// Atomically insert an event AND its thread metadata in a single transaction.
-    #[datastore_span(name = "insert_event_with_thread_metadata", system = "postgresql")]
-    pub async fn insert_event_with_thread_metadata(
-        &self,
-        community_id: CommunityId,
-        event: &nostr::Event,
-        channel_id: Option<Uuid>,
-        thread_meta: Option<event::ThreadMetadataParams<'_>>,
-    ) -> Result<(StoredEvent, bool)> {
-        let result = event::insert_event_with_thread_metadata(
-            &self.pool,
-            community_id,
-            event,
-            channel_id,
-            thread_meta,
-        )
-        .await?;
-        if result.1 {
-            if let Err(e) = insert_mentions(&self.pool, community_id, event, channel_id).await {
-                tracing::warn!(event_id = %event.id, "Failed to insert mentions: {e}");
-            }
-        }
-        Ok(result)
     }
 
     /// Atomically insert a kind:7 reaction event and its reaction row.
@@ -4025,26 +3685,6 @@ impl Db {
         archived_identities::list_archived(&self.pool, community_id).await
     }
 
-    /// Soft-delete NIP-29 discovery events for a channel created by a specific relay pubkey.
-    #[datastore_span(name = "soft_delete_discovery_events", system = "postgresql")]
-    pub async fn soft_delete_discovery_events(
-        &self,
-        community_id: CommunityId,
-        channel_id: Uuid,
-        relay_pubkey: &[u8],
-    ) -> Result<u64> {
-        let result = sqlx::query(
-            "UPDATE events SET deleted_at = NOW() \
-             WHERE community_id = $1 AND channel_id = $2 AND pubkey = $3 AND deleted_at IS NULL AND kind IN (39000, 39001, 39002)",
-        )
-        .bind(community_id.as_uuid())
-        .bind(channel_id)
-        .bind(relay_pubkey)
-        .execute(&self.pool)
-        .await?;
-        Ok(result.rows_affected())
-    }
-
     /// Returns whether the relay-authored NIP-43 snapshot is absent or differs
     /// from the canonical membership rows for `community_id`.
     ///
@@ -4268,75 +3908,6 @@ mod tests {
             .await
             .expect("insert community");
         id
-    }
-
-    #[tokio::test]
-    #[ignore = "requires Postgres"]
-    async fn coordinate_delete_spares_head_newer_than_the_deletion() {
-        use nostr::{EventBuilder, Keys, Kind, Tag, Timestamp};
-
-        let db = setup_db().await;
-        let community = CommunityId::from_uuid(make_community(&db.pool).await);
-        let keys = Keys::generate();
-        let kind = buzz_core::kind::KIND_PROJECT as i32;
-        let d_tag = "stale-tombstone-project";
-        let pubkey = keys.public_key().to_bytes().to_vec();
-        let base = Timestamp::now().as_secs();
-
-        let version = |content: &str, offset: u64| {
-            EventBuilder::new(Kind::Custom(buzz_core::kind::KIND_PROJECT as u16), content)
-                .tags(vec![Tag::parse(["d", d_tag]).expect("d tag")])
-                .custom_created_at(Timestamp::from(base + offset))
-                .sign_with_keys(&keys)
-                .expect("sign project version")
-        };
-
-        for (content, offset) in [("v1", 0), ("v2", 100)] {
-            assert!(
-                db.replace_parameterized_event(community, &version(content, offset), d_tag, None)
-                    .await
-                    .expect("store project version")
-                    .1
-            );
-        }
-
-        // Tombstone timestamped between V1 and V2: it authorizes deleting V1,
-        // never the newer head that replaced it.
-        let stale_deleted = db
-            .soft_delete_by_coordinate(community, kind, &pubkey, d_tag, (base + 50) as i64)
-            .await
-            .expect("stale coordinate delete");
-        assert!(
-            !stale_deleted,
-            "a tombstone older than the live head must delete nothing"
-        );
-
-        let live_content: Option<String> = sqlx::query_scalar(
-            "SELECT content FROM events \
-             WHERE community_id=$1 AND kind=$2 AND pubkey=$3 AND d_tag=$4 AND deleted_at IS NULL",
-        )
-        .bind(community.as_uuid())
-        .bind(kind)
-        .bind(&pubkey)
-        .bind(d_tag)
-        .fetch_optional(&db.pool)
-        .await
-        .expect("read live head");
-        assert_eq!(
-            live_content.as_deref(),
-            Some("v2"),
-            "the newer head must survive a stale tombstone"
-        );
-
-        // A tombstone at or after the head's own timestamp still deletes it.
-        let current_deleted = db
-            .soft_delete_by_coordinate(community, kind, &pubkey, d_tag, (base + 100) as i64)
-            .await
-            .expect("current coordinate delete");
-        assert!(
-            current_deleted,
-            "a tombstone at the head's timestamp must delete it (NIP-09 is at-or-before)"
-        );
     }
 
     #[tokio::test]
